+++ src/meltano/core/elt_context.py
@@ -153,19 +153,21 @@ class ELTContextBuilder:
 
     def context(self, session) -> ELTContext:
         def env_for_plugin(plugin):
+            config_dict = plugin.config_dict(extras=False)
+            config_env = plugin.config_env(extras=False)
             env_struct = {
                 "meltano": {
                     plugin.type.singular: {  # MELTANO_EXTRACTOR_...
                         "name": plugin.name,
                         "namespace": plugin.namespace,
                     },
-                    plugin.type.verb: plugin.config_dict(),  # MELTANO_EXTRACT_...
+                    plugin.type.verb: config_dict,  # MELTANO_EXTRACT_...
                 },
-                **plugin.config_env(),  # TAP_...
+                **config_env,  # TAP_...
             }
-            env_vars = flatten(env_struct, "env_var").items()
+            env_vars = flatten(env_struct, "env_var")
 
-            return {k: str(v) for k, v in env_vars if v is not None}
+            return {k: str(v) for k, v in env_vars.items() if v is not None}
 
         env = {}